+++ packages/frontend/src/navigation/AuthenticatedRoutes.tsx
@@ -3,7 +3,7 @@ import StartPage from '@/pages/Start/StartPage';
 import { ROLE_ADMIN, ROLE_QUANTIFIER } from '@/model/auth';
 import React from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
-import { difference } from 'lodash';
+import difference from 'lodash/difference';
 
 const MyPraisePage = React.lazy(() => import('@/pages/MyPraise/MyPraisePage'));
 const UserDetailsPage = React.lazy(