+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prevent unncessary import of mongoose package by frontend #466
 - Switching eth wallet should result in logging out the user #469
 - Logging out of MetaMask causes EthAccount in navigation to disappear #470
+- Changed lodash imports to single method import #486
 
 ## [0.8.0] - 2022-06-06